feat: add types for user
