+++ news-api/src/entities/users/api/userApi.ts
@@ -1,32 +1,5 @@
 import { baseApi } from '../../../shared';
-
-type Geo = {
-    lat: string;
-    lng: string;
-};
-type Address = {
-    street: string;
-    suite: string;
-    city: string;
-    zipcode: string;
-    geo: Geo;
-};
-type Company = {
-    bs: string;
-    catchPhrase: string;
-    name: string;
-};
-
-export type User = {
-    id: number;
-    name: string;
-    username: string;
-    email: string;
-    company: Company;
-    phone?: string;
-    website?: string;
-    address: Address;
-};
+import type { User } from '../model';
 
 export const userApi = baseApi.injectEndpoints({
     endpoints: (builder) => ({
