+++ news-api/src/entities/users/model/index.ts
@@ -1,1 +1,2 @@
 export { selectAllUsers, usersSelectors, userSlice } from './slice';
+export type { User } from './types';
